requirements.txt
fastapi
uvicorn
redis
debugpy
pyserial
fasthx
pydantic
Jinja2